Thread locking and cancancan auth refactoring for comment threads

+++ app/controllers/comment_threads_controller.rb
@@ -6,13 +6,19 @@ class CommentThreadsController < BoardsBaseController
   def show
     @comments = current_thread.comments.least_recent_first.all.decorate
     @comment = current_thread.comments.build
+    @board = current_board.decorate
     respond_to do |format|
       format.html
     end
   end
 
   def create
-    c_params = comment_thread_params
+    if can?(:create_sticky, current_board)
+      c_params = params.permit(comment_thread: [:sticky], comment: %i[content image])
+    else
+      c_params = params.permit(comment: %i[content image])
+    end
+
     thread = current_board.threads.build(sticky: !!c_params.dig(:comment_thread, :sticky))
     result = CommentBuilder.call(user: current_user,
                                  thread: thread,
@@ -32,24 +38,28 @@ class CommentThreadsController < BoardsBaseController
   end
 
   def unlock
-    if roles?(user: :daemon, board: :moderator)
+    if can?(:lock_thread, current_board)
       current_thread.unlock
       redirect_back_or_to board_url(current_board), notice: "Unlocked thread #{current_thread.id}"
     else
-      redirect_to boards_url, status: :unprocessable_entity, alert: 'Not authorized to unlock thread'
+      redirect_back_or_to boards_url(current_board), alert: 'Not authorized to unlock thread'
     end
   end
 
   def lock
-    if roles?(user: :daemon, board: :moderator)
+    if can?(:lock_thread, current_board)
       current_thread.lock
       redirect_back_or_to board_url(current_board), notice: "Locked thread #{current_thread.id}"
     else
-      redirect_to boards_url, status: :unprocessable_entity, alert: 'Not authorized to lock thread'
+      redirect_back_or_to board_url(current_board), alert: 'Not authorized to lock thread'
     end
   end
 
   def destroy
+    unless can?(:destroy_thread, current_board)
+      redirect_to board_url(current_board), status: :see_other, alert: 'Unauthorized'
+    end
+
     if current_thread.destroy
       redirect_to board_url(current_board), notice: "Thread #{params[:id]} destroyed", status: :see_other
     else
@@ -59,12 +69,8 @@ class CommentThreadsController < BoardsBaseController
 
   private
 
-  def comment_thread_params
-    if roles?(user: :daemon, board: :admin)
-      params.permit(comment_thread: [:sticky], comment: %i[content image])
-    else
-      params.permit(comment: %i[content image])
-    end
+  def current_ability
+    @current_ability ||= ThreadAbility.new(current_user)
   end
 
   def current_thread
